+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Fix] Make sure that the `recent` displays the most recent non-compacted, non-system message.
 - [Fix] Improve the `recent` message display to compensate for aborted transactions.
 - [Fix] Fix `ReferenceError: response is not defined` that occurs when Web UI returns refresh non 200.
+- [Change] Rely on `karafka-core` `>=` `2.2.4` to support lazy loaded custom web producer.
 
 ## 0.7.7 (2023-10-20)
 - [Fix] Remove `thor` as a CLI engine due to breaking changes.